feat(erc20): move contracts dir
