+++ .gitignore
@@ -73,4 +73,6 @@ dependency-graph.png
 # Node.js
 tests/**/node_modules/*
 tests-solidity/**/node_modules/*
-x/erc20/types/contracts/@openzeppelin/*
+
+# OpenZeppelin contracts
+contracts/@openzeppelin/*
