+++ contracts/erc20maliciousdelayed.go
