+++ x/erc20/keeper/msg_server.go
@@ -12,8 +12,8 @@ import (
 
 	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 
+	"github.com/tharsis/evmos/contracts"
 	"github.com/tharsis/evmos/x/erc20/types"
-	"github.com/tharsis/evmos/x/erc20/types/contracts"
 )
 
 var _ types.MsgServer = &Keeper{}
