+++ x/incentives/keeper/keeper_test.go
@@ -38,7 +38,7 @@ import (
 	feemarkettypes "github.com/tharsis/ethermint/x/feemarket/types"
 
 	"github.com/tharsis/evmos/app"
-	"github.com/tharsis/evmos/x/erc20/types/contracts"
+	"github.com/tharsis/evmos/contracts"
 	"github.com/tharsis/evmos/x/incentives/types"
 )
